+++ files/gitlab-cookbooks/gitlab/recipes/gitlab-shell.rb
@@ -24,7 +24,7 @@ gitlab_shell_var_dir = "/var/opt/gitlab/gitlab-shell"
 git_data_directories = node['gitlab']['gitlab-shell']['git_data_directories']
 repositories_storages = node['gitlab']['gitlab-rails']['repositories_storages']
 ssh_dir = File.join(node['gitlab']['user']['home'], ".ssh")
-authorized_keys = File.join(ssh_dir, "authorized_keys")
+authorized_keys = node['gitlab']['gitlab-shell']['auth_file']
 log_directory = node['gitlab']['gitlab-shell']['log_directory']
 hooks_directory = node['gitlab']['gitlab-rails']['gitlab_shell_hooks_path']